BJS2-52694 resolve conflict
