--- src/main/java/school/faang/user_service/client/payment/Currency.java
@@ -1,5 +1,0 @@
-package school.faang.user_service.client.payment;
-
-public enum Currency {
-    USD, EUR
-}
