+++ src/main/java/school/faang/user_service/client/payment/PaymentRequest.java
@@ -2,9 +2,12 @@ package school.faang.user_service.client.payment;
 
 import jakarta.validation.constraints.Min;
 import jakarta.validation.constraints.NotNull;
+import lombok.Builder;
+import school.faang.user_service.common.Currency;
 
 import java.math.BigDecimal;
 
+@Builder
 public record PaymentRequest(
         long paymentNumber,
