+++ src/main/java/school/faang/user_service/controller/UserController.java
@@ -1,12 +1,18 @@
 package school.faang.user_service.controller;
 
 import lombok.RequiredArgsConstructor;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
+import school.faang.user_service.dto.TariffDto;
 import school.faang.user_service.dto.user.UserDto;
-import school.faang.user_service.service.UserService;
+import school.faang.user_service.service.user.UserService;
+
+import java.util.List;
 
 @RestController
 @RequiredArgsConstructor
@@ -17,5 +23,15 @@ public class UserController {
     @GetMapping("/{userId}")
     public UserDto getUser(@PathVariable long userId) {
         return userService.getUser(userId);
+    }
+
+    @PostMapping("/buy-tariff")
+    public ResponseEntity<TariffDto> buyTariff(@RequestBody BuyTariffRequest request) {
+        return ResponseEntity.ok(userService.buyUserTariff(request.tariffDto(), request.id()));
+    }
+
+    @GetMapping("/users")
+    public List<UserDto> getUsers(@RequestBody GetUserRequest request) {
+        return userService.findUsersByFilter(request);
     }
 }
