--- src/main/java/school/faang/user_service/dto/PaymentRequest.java
@@ -1,22 +1,0 @@
-package school.faang.user_service.dto;
-
-import jakarta.validation.constraints.Min;
-import jakarta.validation.constraints.NotNull;
-import lombok.Builder;
-import school.faang.user_service.common.Currency;
-
-import java.math.BigDecimal;
-
-@Builder
-public record PaymentRequest(
-        @NotNull
-        long paymentNumber,
-
-        @Min(1)
-        @NotNull
-        BigDecimal amount,
-
-        @NotNull
-        Currency currency
-) {
-}
