--- src/main/java/school/faang/user_service/dto/PaymentStatus.java
@@ -1,5 +1,0 @@
-package school.faang.user_service.dto;
-
-public enum PaymentStatus {
-    SUCCESS
-}
