+++ src/main/java/school/faang/user_service/exception/BusinessException.java
@@ -1,4 +1,4 @@
-package school.faang.user_service.exceptions;
+package school.faang.user_service.exception;
 
 public class BusinessException extends RuntimeException {
     public BusinessException(String message) {
