+++ src/main/java/school/faang/user_service/exception/DataValidationException.java
@@ -1,4 +1,4 @@
-package school.faang.user_service.exceptions;
+package school.faang.user_service.exception;
 
 public class DataValidationException extends RuntimeException {
     public DataValidationException(String message) {
