+++ src/main/java/school/faang/user_service/exception/PaymentException.java
@@ -1,4 +1,4 @@
-package school.faang.user_service.exceptions;
+package school.faang.user_service.exception;
 
 public class PaymentException extends RuntimeException {
     public PaymentException(String message) {
