+++ src/main/java/school/faang/user_service/exception/PremiumInvalidDataException.java
