+++ src/main/java/school/faang/user_service/exception/ServiceNotAvailableException.java
