--- src/main/java/school/faang/user_service/service/UserService.java
@@ -1,7 +1,0 @@
-package school.faang.user_service.service;
-
-import school.faang.user_service.dto.user.UserDto;
-
-public interface UserService {
-    UserDto getUser(long userId);
-}
