+++ src/main/java/school/faang/user_service/service/user/UserService.java
@@ -1,6 +1,6 @@
 package school.faang.user_service.service.user;
 
-import school.faang.user_service.controller.user.GetUserRequest;
+import school.faang.user_service.controller.GetUserRequest;
 import school.faang.user_service.dto.TariffDto;
 import school.faang.user_service.dto.user.UserDto;
 import school.faang.user_service.entity.User;
@@ -13,4 +13,6 @@ public interface UserService {
     TariffDto buyUserTariff(TariffDto tariffDto, Long userId);
 
     List<UserDto> findUsersByFilter(GetUserRequest request);
+
+    UserDto getUser(long userId);
 }
