+++ src/test/java/school/faang/user_service/service/PremiumServiceTest.java
@@ -4,17 +4,21 @@ import org.jetbrains.annotations.NotNull;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
-import org.mockito.*;
+import org.mockito.ArgumentCaptor;
+import org.mockito.Captor;
+import org.mockito.InjectMocks;
+import org.mockito.Mock;
+import org.mockito.Spy;
 import org.mockito.junit.jupiter.MockitoExtension;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
-import school.faang.user_service.client.PaymentServiceClient;
+import school.faang.user_service.client.payment.PaymentRequest;
+import school.faang.user_service.client.payment.PaymentResponse;
+import school.faang.user_service.client.payment.PaymentServiceFeignClient;
 import school.faang.user_service.common.PaymentStatus;
 import school.faang.user_service.common.PremiumPeriod;
 import school.faang.user_service.config.context.UserContext;
-import school.faang.user_service.dto.PaymentRequest;
 import school.faang.user_service.dto.PremiumDto;
-import school.faang.user_service.dto.response.PaymentResponse;
 import school.faang.user_service.entity.User;
 import school.faang.user_service.entity.premium.Premium;
 import school.faang.user_service.exception.PremiumInvalidDataException;
@@ -27,9 +31,14 @@ import school.faang.user_service.service.impl.PremiumServiceImpl;
 import java.time.LocalDateTime;
 import java.util.Optional;
 
-import static org.junit.jupiter.api.Assertions.*;
+import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.junit.jupiter.api.Assertions.assertNotNull;
+import static org.junit.jupiter.api.Assertions.assertThrows;
 import static org.mockito.ArgumentMatchers.any;
-import static org.mockito.Mockito.*;
+import static org.mockito.Mockito.anyLong;
+import static org.mockito.Mockito.verify;
+import static org.mockito.Mockito.verifyNoInteractions;
+import static org.mockito.Mockito.when;
 
 @ExtendWith(MockitoExtension.class)
 public class PremiumServiceTest {
@@ -39,7 +48,7 @@ public class PremiumServiceTest {
     private static final Boolean USER_IS_NOT_PREMIUM_FLAG = false;
 
     @Mock
-    private PaymentServiceClient paymentServiceClient;
+    private PaymentServiceFeignClient paymentServiceClient;
 
     @Mock
     private PremiumRepository premiumRepository;
@@ -158,6 +167,7 @@ public class PremiumServiceTest {
 
     private @NotNull ResponseEntity<PaymentResponse> getPaymentResponseEntity() {
         return ResponseEntity.ok(new PaymentResponse(
+                1L,
                 PaymentStatus.SUCCESS,
                 1,
                 0L,
