+++ src/test/java/school/faang/user_service/service/tariff/TariffServiceTest.java
@@ -5,10 +5,13 @@ import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.junit.jupiter.MockitoExtension;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.mock.mockito.MockBean;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import school.faang.user_service.BaseTest;
 import school.faang.user_service.client.payment.PaymentResponse;
 import school.faang.user_service.client.payment.PaymentServiceFeignClient;
-import school.faang.user_service.dto.PaymentStatus;
+import school.faang.user_service.common.Currency;
+import school.faang.user_service.common.PaymentStatus;
 import school.faang.user_service.dto.TariffDto;
 import school.faang.user_service.entity.Tariff;
 import school.faang.user_service.entity.User;
@@ -47,13 +50,14 @@ public class TariffServiceTest extends BaseTest {
         TariffDto dto = userServiceProperties.getListAvailableTariffDtos().get(0);
 
         when(paymentServiceFeignClient.sendPayment(any()))
-                .thenReturn(new PaymentResponse(1111L,
-                        PaymentStatus.SUCCESS,
-                        1,
-                        1L,
-                        BigDecimal.valueOf(100),
-                        "USD",
-                        "message"));
+                .thenReturn(new ResponseEntity<>(new PaymentResponse(1111L,
+                                PaymentStatus.SUCCESS,
+                                1,
+                                1L,
+                                BigDecimal.valueOf(100),
+                                Currency.USD,
+                                "message"), HttpStatus.OK)
+                        );
 
         Tariff.TariffBuilder tariff = Tariff.builder()
                 .plan("super-user")
